// frontend/src/client/types.gen.ts
export type MachinePublic = {
    name: string;
    description?: (string | null);
    kep_tag_prefix?: (string | null);
    display_order?: number;
    machine_type?: (string | null);
    id: string;
    unit_id: string;
};